+++ Forge/src/main/java/com/tristankechlo/random_mob_sizes/ForgePlatformHelper.java
@@ -1,5 +1,6 @@
 package com.tristankechlo.random_mob_sizes;
 
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.loading.FMLPaths;
 
 import java.nio.file.Path;
@@ -9,6 +10,11 @@ public class ForgePlatformHelper implements IPlatformHelper {
     @Override
     public Path getConfigDirectory() {
         return FMLPaths.CONFIGDIR.get();
+    }
+
+    @Override
+    public boolean isModLoaded(String modid) {
+        return ModList.get().isLoaded(modid);
     }
 
 }